Inctive to Active for version

+++ GSC.Respository/Project/Design/IStudyVersionRepository.cs
@@ -13,7 +13,7 @@ namespace GSC.Respository.Project.Design
         List<StudyVersionGridDto> GetStudyVersion(int ProjectDesignId);
         string Duplicate(StudyVersion objSave);
         void UpdateVisitStatus(StudyVersion studyVersion);
-        double GetVersionNumber(int projectId, bool isMonir);
+        decimal GetVersionNumber(int projectId, bool isMonir);
         List<DropDownDto> GetVersionDropDown(int projectId);
         bool IsOnTrialByProjectDesing(int projectDesignId);
         void UpdateGoLive(int projectId);
